Изменен тип элементов(int -> double)

+++ Matrics_Functions/Matrics_Functions/LongVector.h
@@ -3,7 +3,7 @@
 namespace Toplitz_Matrix
 {
 	typedef int TIndex;
-	typedef int TElement;
+	typedef double TElement;
 	class Vector
 	{
